[debian] FIX: Warning about missing dependency `python3-apt`.
When running the playbook on debian based systems the `python3-apt`
package is necessary to install other packages and a warning was
being produced that this dependency was being automatically installed.

In order to remove the warning and streamline the testing procedure
we install the package when building the Docker image used for testing.